# Conflicts:
#	GameServerAdmin/GameServerAdmin/Common/Exceptions/AppException.cs
#	GameServerAdmin/GameServerAdmin/Common/Exceptions/InvalidStateException.cs

+++ GW2_Client/Assets/Scripts/Scene/GameScene.cs
@@ -16,7 +16,7 @@ public class GameScene : BaseScene
         //_resourceService.Instantiate("Player/Police");
         // TODO : 하드코딩된 RoomId 변경
         C_ENTER_GAME enterGamePkt = new C_ENTER_GAME();
-        enterGamePkt.PlayerIndex = (ulong)network.GetNetworkId();
+        enterGamePkt.PlayerIndex = network.GetNetworkId();
         enterGamePkt.RoomId = 0;
         network.Send(enterGamePkt);
     }
